+++ src/handlers/menu.ts
@@ -4,7 +4,8 @@ import { isScribeAdmin } from "../services/adminAccess";
 import { isPlayerAutoEnabled } from "./auto";
 
 export async function showMenu(ctx: any) {
-  await ctx.reply("☰ Меню", { reply_markup: buildMenuReplyKeyboard({ canSeeStats: await isScribeAdmin(ctx.from?.id) }) });
+  const canSeeScribeTools = await isScribeAdmin(ctx.from?.id);
+  await ctx.reply("☰ Меню", { reply_markup: buildMenuReplyKeyboard({ canSeeStats: canSeeScribeTools, canSeeChat: canSeeScribeTools }) });
 }
 
 export async function showMainKeyboard(ctx: any) {